scanner: fix compiler warning
src/xkbcomp/scanner.c:158:17: warning: comparison of constant -1 with expression of type 'enum yytokentype' is always true
      [-Wtautological-constant-out-of-range-compare]
        if (tok != -1) return tok;
            ~~~ ^  ~~

+++ src/xkbcomp/scanner.c
@@ -155,7 +155,7 @@ skip_more_whitespace_and_comments:
 
         /* Keyword. */
         tok = keyword_to_token(s->buf);
-        if (tok != -1) return tok;
+        if ((int) tok != -1) return tok;
 
         yylval->str = strdup(s->buf);
         if (!yylval->str)
